Fixed type hinting

+++ carton/data.py
@@ -10,10 +10,11 @@ import numpy as np
 
 
 def describe_series(
-    s: Sequence, r: int = 2, qs: Optional[Sequence[int]] = None
+    s: Sequence, r: int = 2, qs: Optional[Sequence[float]] = None
 ) -> dict:
     if not qs:
         qs = [10, 25, 50, 75, 90, 95, 99, 99.9, 99.99]
+
     info = {
         "size": len(s),
         "mode": statistics.mode(s),
